requirements/base.txt
django>=2.0.1,<3.0.0
django-environ==0.4.4
djangorestframework>=3.0.0,<4.0.0
django-rest-auth>=0.9.2,<1.0.0
djangorestframework-jwt>=1.11.0,<2.0.0
django-allauth>=0.33.0,<1.0.0
django-model-utils>=3.1.1,<4.0.0
Pillow>=5.0.0,<6.0.0
